refactor(rendering)!: isolate pipeline unsafe into SubtreeArena + fix pre-existing cyclic-edge UB
Phase 2 keystone (architectural). Extract 100% of the pipeline's unsafe re-entrant
layout machinery out of the 5622-LOC owner.rs god-file into one auditable module
`pipeline/owner/subtree_arena.rs` behind a SAFE API. owner.rs → owner/mod.rs (git mv);
`owner/` now contains ZERO executable unsafe.

Moved together (they cross-call, so partial extraction would leak unsafe back):
NodePtr + `unsafe impl Send/Sync`, SubtreeBorrows → SubtreeArena, LayoutCycleGuard,
ensure_stack, and all THREE mutually-recursive unsafe walks (layout / box-intrinsic /
sliver-layout). Safe surface: SubtreeArena::from_tree / layout_child /
layout_sliver_child / box_intrinsic; the `unsafe fn`s are module-private. NodePtr
cannot outlive the borrow (PhantomData<&'tree mut ()>). The Send+Sync on the callback
is load-bearing per ADR-0002 (data plane Send-ready for future parallel layout), so the
unsafe is isolated, not deleted.

FIXES A PRE-EXISTING MIRI-CONFIRMED UB (was latent in main): on a cyclic child edge the
child-seeding shared read `&*child_ptr.0` aliased the parent's live `&mut RenderNode`
(use-after-pop). The `// SAFETY:` premise "acyclicity ⇒ parent ≠ child" was false —
LayoutCycleGuard blocks the 2nd `&mut` reborrow but not the earlier shared read. The
new adversarial Miri tests surfaced it (old pipeline-level Miri missed it). Fixed by an
`is_in_flight` skip before every child shared read (box + sliver, seed + commit, 4 sites):
behavior-preserving on acyclic trees, clean LayoutCycle error on cyclic edges.

Verification (independently re-run): cargo +nightly miri test u21_layout_cycle 4/4 under
BOTH Stacked Borrows AND Tree Borrows (the previously-UB reentry test now passes);
u20_layout_dirty_root 11/11; nextest 935/935; clippy -D warnings + check --workspace +
port-check 20/20 clean. SAFETY-GATE PASS (unsafe-auditor + systems-perf-lead co-sign).

Perf: extraction is zero-regression (no new alloc/dyn/indirection). The per-child
is_in_flight Mutex::lock (~25ns uncontended, dominated by perform_layout) is the cost of
the Send+Sync bound; a Cell<usize> fast-path is a tracked follow-up.

Out of scope (follow-up Phase-2 units): DirtyTracker split, phase impl-split, capability-
trait collapse, RenderAlign/RenderProxySliver bases.
