+++ crates/flui-rendering/src/pipeline/owner/mod.rs
@@ -6,6 +6,9 @@
 //! `run_paint` on `<Idle>` is a compile error -- see the `compile_fail`
 //! doctest at the end of `pipeline/phase.rs`.
 
+mod subtree_arena;
+use subtree_arena::{SubtreeArena, child_flex_from_seeds, ensure_stack};
+
 use std::{
     marker::PhantomData,
     sync::atomic::{AtomicBool, AtomicU64, Ordering},
@@ -18,7 +21,6 @@ use flui_layer::{
 };
 use flui_painting::DisplayList;
 use flui_types::{Offset, Size};
-use parking_lot::Mutex;
 #[cfg(any(test, feature = "testing"))]
 use rustc_hash::FxHashMap;
 use rustc_hash::FxHashSet;
@@ -29,15 +31,8 @@ use crate::testing::parent_data::ParentDataSeed;
 use crate::{
     constraints::{BoxConstraints, SliverConstraints, SliverGeometry},
     context::{FragmentClip, FragmentOp, FragmentRecorder},
-    protocol::{
-        BoxProtocol, MainAxisPosition, Protocol, SliverProtocol,
-        box_protocol::{
-            ActualBaselineChildCallback, BoxLayoutCtxErased, LayoutChildCallback,
-            SliverLayoutChildCallback,
-        },
-        sliver_protocol::{SliverChildLayoutCallback, SliverLayoutCtxErased},
-    },
-    storage::{RenderEntry, RenderNode, RenderTree},
+    protocol::{BoxProtocol, MainAxisPosition, SliverProtocol},
+    storage::{RenderNode, RenderTree},
 };
 
 use super::{
@@ -2399,67 +2394,35 @@ impl PipelineOwner<Layout> {
         id: RenderId,
         constraints: BoxConstraints,
     ) -> crate::error::RenderResult<Size> {
-        // Step 1: collect every id in the subtree rooted at `id`
-        // (DFS pre-order). Empty result = id not in tree.
-        let subtree_ids = self.render_tree.collect_subtree_ids(id);
-        if subtree_ids.is_empty() {
-            return Err(crate::error::RenderError::NodeNotFound(id));
-        }
-
-        // Step 2: pre-acquire N disjoint &mut RenderNode borrows on
-        // every subtree slot in ONE function scope. `get_subtree_mut`
-        // returns None on (a) duplicate ids or (b) missing slab slots.
-        // Per `collect_subtree_ids`'s PR #145 visited-set fix the
-        // returned id list is GUARANTEED deduplicated, so case (a) is
-        // unreachable here — None can only mean a slab slot
-        // disappeared between collect and acquire (a race-condition
-        // shape that doesn't occur with &mut self access; defensive
-        // fallback only). NodeNotFound is the most accurate variant
-        // for that residual case.
-        let node_refs = self
-            .render_tree
-            .get_subtree_mut(&subtree_ids)
-            .ok_or(crate::error::RenderError::NodeNotFound(id))?;
-
-        // Step 3: wrap the borrows in a SubtreeBorrows index for O(1)
-        // by-id lookup. The HashMap holds NodePtr (raw alias of the
-        // &mut RenderNode borrows just acquired) so the recursive walk
-        // can reborrow one slot at a time without re-entering the
-        // tree's slab borrow.
-        let borrows = SubtreeBorrows::new(
-            &subtree_ids,
-            node_refs,
+        // Steps 1–3: collect subtree ids, pre-acquire disjoint &mut borrows,
+        // and wrap them in a SubtreeArena for O(1) by-id lookup during the
+        // recursive walk.  The unsafe aliasing machinery lives entirely inside
+        // `subtree_arena::SubtreeArena`; this call site is safe.
+        let arena = SubtreeArena::from_tree(
+            &mut self.render_tree,
+            id,
             #[cfg(any(test, feature = "testing"))]
             &self.parent_data_seeds,
-        );
+        )?;
 
-        // Step 4: recursive walk via index-into-pre-acquired-pool. No
-        // &mut RenderTree appears inside the callback chain — only
-        // per-slot NodePtr reborrows on distinct slab slots, sound
-        // under Stacked / Tree Borrows.
-        //
-        // SAFETY: `borrows` is alive for the entire walk; each
-        // `layout_subtree_borrowed` call reborrows exactly one slot via
-        // its NodePtr; distinct call levels reborrow distinct slots
-        // (parent ≠ child in a well-formed tree).
-        let result = unsafe { layout_subtree_borrowed(&borrows, id, constraints) };
+        // Step 4: recursive walk via the safe arena API.  Internally the arena
+        // reborrows one NodePtr per call level (distinct slots; parent ≠ child
+        // enforced by LayoutCycleGuard), but no `unsafe` appears here.
+        let result = arena.layout_child(id, constraints);
 
         // Step 5: drain on-demand child builds AND removals the walk recorded
-        // (the re-entrant build contract's v1 next-frame backend, ADR-0003
-        // Decision 2 + U3c D2). The walk's tree borrows were frozen, so a lazy
-        // sliver that requested a build or removal parked it in the pending sinks
-        // rather than mutating the tree. Take both (owned), then DROP `borrows`
-        // to release the `&mut RenderTree` subtree borrow before touching
-        // `&mut self`.
+        // (re-entrant build contract v1, ADR-0003 Decision 2 + U3c D2).
+        // Take both (owned), then DROP `arena` to release the &mut RenderTree
+        // subtree borrow before touching `&mut self`.
         //
-        // D3 ordering: Remove → Insert → Update. Removes are drained first so
+        // D3 ordering: Remove → Insert → Update.  Removes are drained first so
         // that an off-band child evicted this pass does not collide with the
         // Insert that replaces it in the same batch.
-        let pending_removes = borrows.take_pending_removes();
-        let pending_builds = borrows.take_pending_builds();
-        drop(borrows);
+        let pending_removes = arena.take_pending_removes();
+        let pending_builds = arena.take_pending_builds();
+        drop(arena);
 
-        // Apply removes first (D3 ordering). Each entry is `(parent, child)`:
+        // Apply removes first (D3 ordering).  Each entry is `(parent, child)`:
         // the parent is the sliver's own node_id (tagged at push time in
         // ErasedSliverLayoutCtx::dispose_box_child), NOT the walk root `id`.
         // Using `id` here would misdirect `mark_needs_layout` to the viewport
@@ -2480,1072 +2443,6 @@ impl PipelineOwner<Layout> {
 
         result
     }
-}
-
-// ============================================================================
-// PR-A1b3 U20.1 — pre-acquired-subtree layout helper (Miri-clean)
-// ============================================================================
-
-/// `Send + Sync` raw-pointer alias of a single `&mut RenderNode` borrow
-/// held in [`SubtreeBorrows`].
-///
-/// Each `NodePtr` in [`SubtreeBorrows::by_id`] is derived from one of
-/// the N disjoint `&mut RenderNode` references returned by
-/// [`RenderTree::get_subtree_mut`]. The pointer is stable for the
-/// lifetime of the `SubtreeBorrows` instance because the underlying
-/// `&mut RenderTree` is held by the caller (`PipelineOwner` while
-/// `layout_dirty_root` runs) and the slab's slot allocation is
-/// position-stable (no moves during the borrow window).
-///
-/// The wrapper is `Copy` so the layout-child closure can capture the
-/// pointer by value without `Arc` ceremony. `Send + Sync` is declared
-/// because [`LayoutChildCallback`] inherits those bounds from
-/// `BoxLayoutCtxErased: Send + Sync` (U19 design). Single-thread
-/// access is enforced at the [`SubtreeBorrows::check_thread`] entry.
-#[derive(Clone, Copy)]
-struct NodePtr(*mut RenderNode);
-
-// SAFETY: the raw pointer is just an address; the load-bearing borrow
-// is the `&mut RenderNode` returned by `get_subtree_mut` that this
-// pointer aliases. Cross-thread reborrow is rejected by
-// [`SubtreeBorrows::check_thread`] before any deref.
-unsafe impl Send for NodePtr {}
-// SAFETY: same as Send.
-unsafe impl Sync for NodePtr {}
-
-/// Pre-acquired set of N disjoint `&mut RenderNode` borrows on a
-/// subtree, indexed by [`RenderId`] for O(1) lookup.
-///
-/// **D-block PR-A1b3 U20.1:** replaces the prior `TreePtr` +
-/// recursive-tree-reborrow scheme (PR #144) that surfaced as latent
-/// Stacked / Tree Borrows UB. The new scheme acquires ALL subtree
-/// `&mut RenderNode` borrows in ONE call to
-/// [`RenderTree::get_subtree_mut`] (single `&mut Slab` reborrow scope),
-/// stores raw aliases in this map, and lets the recursive walk reborrow
-/// one slot at a time per call level. No `&mut RenderTree` ever appears
-/// inside the layout-child callback chain — eliminates the UB.
-///
-/// # Lifetime
-///
-/// `'tree` ties `SubtreeBorrows` to the source `&mut RenderTree`
-/// borrow's lifetime via `PhantomData<&'tree mut ()>`. Constructed via
-/// [`Self::new`] from a `Vec<&'tree mut RenderNode>` (the output of
-/// `get_subtree_mut`); the references are immediately converted to
-/// raw pointers and aggregated by id. The `&mut RenderTree` source
-/// borrow keeps the slab's slots position-stable for the lifetime of
-/// every aliased `NodePtr`.
-///
-/// # Thread affinity
-///
-/// `SubtreeBorrows` records the constructing thread's `ThreadId` and
-/// checks it on every [`Self::get`] call. The check survives even
-/// though [`NodePtr`] declares `Send + Sync` — the auto-trait bound
-/// is mechanically required to satisfy
-/// `LayoutChildCallback: Send + Sync` (inherited from
-/// `BoxLayoutCtxErased`), but at the call site we panic loudly on
-/// cross-thread access instead of corrupting the slab silently.
-/// Cheap: one `ThreadId::eq` per lookup.
-struct SubtreeBorrows<'tree> {
-    by_id: std::collections::HashMap<RenderId, NodePtr>,
-    #[cfg(any(test, feature = "testing"))]
-    parent_data_seeds: FxHashMap<RenderId, ParentDataSeed>,
-    /// Set of ids whose layout is currently in flight at some recursion
-    /// level above the current call. Insert on layout entry, remove on
-    /// drop (RAII via [`LayoutCycleGuard`]). Re-entry on a member id
-    /// surfaces as [`crate::error::RenderError::LayoutCycle`] — closes
-    /// the U21 cycle-detection blocker (companion memo D6).
-    ///
-    /// Wrapped in `parking_lot::Mutex` because the layout-child closure
-    /// requires `&SubtreeBorrows: Send + Sync` (inherited from
-    /// `BoxLayoutCtxErased`). Uncontended `parking_lot::Mutex` acquire
-    /// is ~10 ns — negligible vs `perform_layout` cost. The cross-
-    /// thread closure-smuggle attack vector is independently rejected
-    /// by [`Self::check_thread`], so the Mutex serves only as the
-    /// shared-mutability cell, not as actual cross-thread sync.
-    currently_laying_out: Mutex<FxHashSet<RenderId>>,
-    /// On-demand child builds requested during this walk that the frozen
-    /// mid-pass borrows could not insert synchronously — the re-entrant build
-    /// contract's v1 next-frame backend (ADR-0003 Decision 2). `layout_dirty_root`
-    /// drains this into the deferred-mutation queue after the walk releases its
-    /// borrows. `Mutex` for the same reason as `currently_laying_out`: the
-    /// layout-child closure requires `&SubtreeBorrows: Send + Sync`. Empty unless
-    /// a lazy sliver requests a not-yet-built child.
-    pending_builds: Mutex<Vec<crate::protocol::sliver_protocol::PendingBuild>>,
-    /// Symmetric remove sink (U3c D2): `(parent, child)` pairs of children the
-    /// consumer wants evicted from the tree.  The `parent` is the sliver's own
-    /// `node_id` — **not** the walk root `id` passed to `layout_dirty_root`.
-    /// For a real `viewport → sliver → lazy → child` chain the walk root is the
-    /// viewport, but `defer_remove` / `mark_needs_layout` must target the lazy
-    /// sliver so it reflows after its child list changes.
-    ///
-    /// Drained before `pending_builds` in `layout_dirty_root`
-    /// (Remove → Insert ordering, D3), post-drop of the subtree borrows, so no
-    /// aliased `NodePtr` is live when the `defer_remove` calls touch `&mut self`.
-    /// `Mutex` for the same reason as `pending_builds`.
-    pending_removes: Mutex<Vec<(flui_foundation::RenderId, flui_foundation::RenderId)>>,
-    owner_thread: std::thread::ThreadId,
-    _lifetime: std::marker::PhantomData<&'tree mut ()>,
-}
-
-impl<'tree> SubtreeBorrows<'tree> {
-    /// Constructs a `SubtreeBorrows` from the output of
-    /// [`RenderTree::collect_subtree_ids`] paired with the matching
-    /// output of [`RenderTree::get_subtree_mut`].
-    ///
-    /// Precondition: `ids.len() == refs.len()` and each
-    /// `ids[i]` corresponds to `refs[i]` (in order). Caller must
-    /// satisfy this — currently the only caller is
-    /// [`PipelineOwner::layout_dirty_root`] which feeds the two
-    /// methods' outputs directly to this ctor.
-    fn new(
-        ids: &[RenderId],
-        refs: Vec<&'tree mut RenderNode>,
-        #[cfg(any(test, feature = "testing"))] all_seeds: &FxHashMap<RenderId, ParentDataSeed>,
-    ) -> Self {
-        debug_assert_eq!(
-            ids.len(),
-            refs.len(),
-            "SubtreeBorrows::new precondition violated: ids and refs \
-             must have the same length",
-        );
-        let owner_thread = std::thread::current().id();
-        let mut by_id = std::collections::HashMap::with_capacity(ids.len());
-        for (&id, r) in ids.iter().zip(refs) {
-            by_id.insert(id, NodePtr(r as *mut RenderNode));
-        }
-        #[cfg(any(test, feature = "testing"))]
-        let parent_data_seeds = ids
-            .iter()
-            .filter_map(|id| all_seeds.get(id).map(|seed| (*id, seed.clone())))
-            .collect();
-        Self {
-            by_id,
-            #[cfg(any(test, feature = "testing"))]
-            parent_data_seeds,
-            // Pre-sized to subtree size — at most `ids.len()` entries
-            // can be in-flight concurrently (the recursive walk
-            // descends linearly through one path at a time).
-            currently_laying_out: Mutex::new(FxHashSet::with_capacity_and_hasher(
-                ids.len(),
-                Default::default(),
-            )),
-            pending_builds: Mutex::new(Vec::new()),
-            pending_removes: Mutex::new(Vec::new()),
-            owner_thread,
-            _lifetime: std::marker::PhantomData,
-        }
-    }
-
-    #[cfg(any(test, feature = "testing"))]
-    fn seed_child_parent_data(
-        &self,
-        child_id: RenderId,
-        slot: &mut Option<Box<dyn crate::parent_data::ParentData>>,
-    ) {
-        if let Some(seed) = self.parent_data_seeds.get(&child_id) {
-            *slot = Some(seed.to_box());
-        }
-    }
-
-    /// Panics if the calling thread is not the constructing thread.
-    /// Called by [`Self::get`] before returning any [`NodePtr`].
-    #[inline]
-    fn check_thread(&self) {
-        let current = std::thread::current().id();
-        if current != self.owner_thread {
-            panic!(
-                "SubtreeBorrows accessed from non-owner thread: \
-                 owner = {:?}, current = {:?}. The U20 layout walk \
-                 requires the layout_child callback to fire on the \
-                 same thread as PipelineOwner::layout_dirty_root \
-                 (the pipeline phase holds &mut self synchronously). \
-                 User RenderBox::perform_layout body must not spawn \
-                 ctx.layout_child(...) calls to other threads — the \
-                 underlying RenderTree slab is not Sync.",
-                self.owner_thread, current,
-            );
-        }
-    }
-
-    /// Returns the [`NodePtr`] for `id` if present, panicking
-    /// (via [`Self::check_thread`]) on cross-thread access.
-    #[inline]
-    fn get(&self, id: RenderId) -> Option<NodePtr> {
-        self.check_thread();
-        self.by_id.get(&id).copied()
-    }
-
-    /// Takes the on-demand child builds recorded during this walk, leaving the
-    /// sink empty. Called by [`PipelineOwner::layout_dirty_root`] once the walk
-    /// has returned, so the requests can be enqueued on the deferred-mutation
-    /// queue (which needs `&mut PipelineOwner`). Touches only the sink — never a
-    /// [`NodePtr`] — so it does not interact with the raw-pointer aliasing.
-    fn take_pending_builds(&self) -> Vec<crate::protocol::sliver_protocol::PendingBuild> {
-        std::mem::take(&mut *self.pending_builds.lock())
-    }
-
-    /// Takes the deferred child removals recorded during this walk.
-    /// Returns `(parent, child)` pairs — the parent is the sliver's `node_id`,
-    /// not the walk root — so `defer_remove` targets the correct ancestor.
-    /// Symmetric to [`Self::take_pending_builds`]; called in `layout_dirty_root`
-    /// BEFORE `take_pending_builds` (Remove → Insert ordering, D3) and AFTER
-    /// `drop(borrows)` so no `NodePtr` alias is live when the removes are applied.
-    fn take_pending_removes(&self) -> Vec<(flui_foundation::RenderId, flui_foundation::RenderId)> {
-        std::mem::take(&mut *self.pending_removes.lock())
-    }
-}
-
-// ============================================================================
-// PR-A1 U21 — RAII layout-cycle guard
-// ============================================================================
-
-/// RAII guard that registers `id` in [`SubtreeBorrows::currently_laying_out`]
-/// on construction and unregisters on drop.
-///
-/// **D-block PR-A1 U21 (companion memo D6):** detects re-entry into a
-/// node's `layout_subtree_borrowed` call (the situation where a user
-/// `perform_layout` body calls `ctx.layout_child` for an ancestor id
-/// whose layout is already in flight up the stack). On collision the
-/// constructor returns [`crate::error::RenderError::LayoutCycle`]
-/// instead of attempting a second [`NodePtr`] reborrow (which would be
-/// UB under aliasing rules — the same slot's Unique tag is live up the
-/// recursion stack).
-///
-/// The guard's `Drop` impl unconditionally removes `id` from the set,
-/// even on unwind (Rust's drop semantics guarantee this for any
-/// `Drop`-implementing value going out of scope). Combined with the
-/// `catch_unwind` wrapper around `perform_layout_raw` in the non-leaf
-/// path, this means the cycle set stays consistent across frames: a
-/// panicking widget's id is cleared, the next frame's walk does not
-/// see it as in-flight.
-struct LayoutCycleGuard<'b, 'tree> {
-    borrows: &'b SubtreeBorrows<'tree>,
-    id: RenderId,
-}
-
-impl<'b, 'tree> LayoutCycleGuard<'b, 'tree> {
-    /// Registers `id` as currently-laying-out. Returns
-    /// `Err(RenderError::LayoutCycle(id))` if `id` is already
-    /// registered — caller must propagate immediately.
-    fn enter(borrows: &'b SubtreeBorrows<'tree>, id: RenderId) -> crate::error::RenderResult<Self> {
-        // check_thread here so the diagnostic surfaces at the cycle-
-        // guard layer too (covers callers that bypass `get`).
-        borrows.check_thread();
-        let mut set = borrows.currently_laying_out.lock();
-        if !set.insert(id) {
-            // Debug-level: the layout-child callback in
-            // `layout_subtree_borrowed` already logs the propagated
-            // Err at tracing::error when it collapses descendant Err
-            // to Size::ZERO. Logging here at error too would produce
-            // 2 log lines per cycle event (PR #146 Copilot review
-            // comment 3294315141). The API-boundary error log is the
-            // user-facing one; this debug-level log retains the
-            // collision-point diagnostic for tracing.
-            tracing::debug!(
-                ?id,
-                "layout_subtree_borrowed: layout cycle detected — id is \
-                     already in flight at a parent call level; returning \
-                     RenderError::LayoutCycle(id)",
-            );
-            return Err(crate::error::RenderError::layout_cycle(id));
-        }
-        // Lock drops here — set is held only for the insert.
-        Ok(Self { borrows, id })
-    }
-}
-
-impl<'b, 'tree> Drop for LayoutCycleGuard<'b, 'tree> {
-    fn drop(&mut self) {
-        // Unconditional remove — runs on every exit path including
-        // unwind. Cycle set stays consistent for the next frame.
-        // `Mutex::lock` is panic-safe (no poisoning in parking_lot).
-        self.borrows.currently_laying_out.lock().remove(&self.id);
-    }
-}
-
-/// Grows the stack ahead of each pipeline-walk recursion level so
-/// arbitrarily deep render trees cannot overflow the fixed OS stack
-/// (the Windows main thread gets 1 MiB by default; a ~1000-level
-/// single-child chain blew it in the `layout/deep/1000` bench, and a
-/// production tree of that depth would crash the app identically).
-///
-/// Same discipline as rustc's `ensure_sufficient_stack`: when fewer
-/// than the red-zone bytes remain, the continuation runs on a fresh
-/// heap-allocated stack segment. Cost on the hot path is one
-/// stack-pointer probe per recursion level (sub-ns next to the
-/// per-node layout/paint work).
-///
-/// Falls back to a direct call under miri (psm's stack-switching
-/// assembly cannot be interpreted) and on wasm32 (no stack switching;
-/// the dependency is compiled out in Cargo.toml) — those environments
-/// keep plain recursion and its pre-existing depth limits.
-#[inline]
-fn ensure_stack<R>(f: impl FnOnce() -> R) -> R {
-    #[cfg(any(miri, target_arch = "wasm32"))]
-    {
-        f()
-    }
-    #[cfg(not(any(miri, target_arch = "wasm32")))]
-    {
-        // 128 KiB red zone: covers the deepest single-level frame
-        // chain between two probes (driver frame + typed ctx + the
-        // render object's own perform_layout/paint locals). 2 MiB
-        // segments amortize one allocation across many levels.
-        stacker::maybe_grow(128 * 1024, 2 * 1024 * 1024, f)
-    }
-}
-
-/// Recursive helper for [`PipelineOwner::layout_dirty_root`].
-///
-/// Reborrows one [`NodePtr`] from the pre-acquired [`SubtreeBorrows`]
-/// at each call level, drives `perform_layout_raw` against a typed
-/// `BoxLayoutCtx`, and recurses via a closure that captures
-/// `&SubtreeBorrows` (Sync via [`NodePtr`]'s `unsafe impl`). Distinct
-/// call levels reborrow distinct slab slots (parent ≠ child) — no
-/// aliasing.
-///
-/// # Safety
-///
-/// Caller must guarantee:
-///
-/// 1. `borrows` is alive for the entire duration of this call AND
-///    every recursive call this helper triggers via the callback. The
-///    [`PipelineOwner::layout_dirty_root`] flow constructs
-///    `SubtreeBorrows` on the caller's stack and only invokes this
-///    helper while the binding is live.
-/// 2. At any moment, no two concurrent reborrows of the SAME
-///    [`NodePtr`] exist. Sequential call levels (parent → child →
-///    grandchild) reborrow DIFFERENT slots — preserved by the U21
-///    `LayoutCycleGuard` (returns
-///    [`crate::error::RenderError::LayoutCycle`] on re-entry into
-///    a slot already in flight up the stack).
-unsafe fn layout_subtree_borrowed<'tree>(
-    borrows: &SubtreeBorrows<'tree>,
-    id: RenderId,
-    constraints: BoxConstraints,
-) -> crate::error::RenderResult<Size> {
-    ensure_stack(|| {
-        // SAFETY: identical contract, forwarded verbatim from this
-        // wrapper's own `# Safety` section; the stack-growth wrapper
-        // only relocates which memory the frames live in, never their
-        // borrow structure, lifetimes, or drop order.
-        unsafe { layout_subtree_borrowed_impl(borrows, id, constraints) }
-    })
-}
-
-/// Body of [`layout_subtree_borrowed`]; split out so every recursion
-/// level enters through the [`ensure_stack`] probe.
-///
-/// # Safety
-///
-/// Same contract as [`layout_subtree_borrowed`].
-unsafe fn layout_subtree_borrowed_impl<'tree>(
-    borrows: &SubtreeBorrows<'tree>,
-    id: RenderId,
-    constraints: BoxConstraints,
-) -> crate::error::RenderResult<Size> {
-    // U21 cycle guard: register `id` in currently_laying_out *before*
-    // any NodePtr reborrow. Drop on every exit path (RAII) — set stays
-    // consistent across panics via the catch_unwind in the non-leaf
-    // path below + Rust's drop-on-unwind discipline. Re-entry returns
-    // RenderError::LayoutCycle(id) immediately, skipping the reborrow
-    // attempt entirely (avoids the otherwise-UB second-Unique-tag on
-    // an in-flight slot).
-    let _cycle_guard = LayoutCycleGuard::enter(borrows, id)?;
-
-    // Resolve id → NodePtr. Cross-thread access panics inside `get`.
-    let NodePtr(node_ptr) = match borrows.get(id) {
-        Some(np) => np,
-        None => return Err(crate::error::RenderError::NodeNotFound(id)),
-    };
-
-    // SAFETY: `node_ptr` aliases a `&mut RenderNode` that
-    // `SubtreeBorrows` holds disjointly from every other slot it
-    // covers. The reborrow below is the FIRST and ONLY reborrow of
-    // `id`'s slot in this call frame; the recursive callback below
-    // reborrows DIFFERENT slots (child ≠ parent by tree acyclicity).
-    // Distinct slot reborrows have independent Unique tags under
-    // Stacked / Tree Borrows — no aliasing.
-    let node_ref: &mut RenderNode = unsafe { &mut *node_ptr };
-
-    // Extract typed RenderEntry<BoxProtocol> + snapshot child_ids in
-    // one scope. Distinguish missing-from-tree (handled before via
-    // borrows.get) from present-but-wrong-protocol.
-    let node_protocol = node_ref.protocol_name();
-    let entry: &mut RenderEntry<BoxProtocol> = match node_ref.as_box_mut() {
-        Some(e) => e,
-        None => {
-            return Err(crate::error::RenderError::ProtocolMismatch {
-                node_protocol,
-                constraints_protocol: "Box",
-            });
-        }
-    };
-    let child_ids: Vec<RenderId> = entry.links().children().to_vec();
-
-    // Short-circuit clean children: if NEEDS_LAYOUT is not set AND
-    // constraints match the cached value, skip layout entirely.
-    // (Flutter rendering/object.dart:2852: early return before recurse)
-    if !entry.needs_layout() && entry.state().has_constraints(&constraints) {
-        // If constraints match and layout is clean, geometry MUST be present
-        // from prior layout. This is a structural invariant.
-        if let Some(geometry) = entry.state().geometry() {
-            return Ok(geometry);
-        }
-        // If not, log and continue with layout (should never happen).
-        tracing::warn!(
-            node_id = ?id,
-            "layout short-circuit: clean constraints cache but missing geometry; \
-             proceeding with layout (invariant violation)"
-        );
-    }
-
-    // Leaf path: delegate to layout_leaf_only (constructs typed
-    // BoxLayoutCtx<Leaf> + catch_unwind + state update + relayout-
-    // boundary bootstrap). Same code as the U18/U19 leaf bridge tests
-    // exercise.
-    if child_ids.is_empty() {
-        return entry.layout_leaf_only(constraints);
-    }
-
-    // Non-leaf path: build ChildState backing vec + descendant-error
-    // flag + recursive callback. Same shape as the prior U20 version,
-    // but the callback recurses into `layout_subtree_borrowed` which
-    // uses pre-acquired NodePtrs instead of fresh tree reborrows.
-    let mut child_states: Vec<crate::protocol::ErasedChildState> = child_ids
-        .iter()
-        .map(|&cid| crate::protocol::ErasedChildState::new(cid))
-        .collect();
-
-    // Seed each ChildState.offset from the child's persisted
-    // RenderState.offset. A parent that does not re-position a child
-    // during this walk must preserve the child's prior offset
-    // (Flutter parity: BoxParentData.offset persists until
-    // positionChild overwrites it) — without the seed, the
-    // post-layout commit below would silently reset unpositioned
-    // children to Offset::ZERO. Box parents can now host both Box and
-    // Sliver children, so seed through RenderNode's protocol-generic
-    // offset view.
-    for cs in &mut child_states {
-        if let Some(child_ptr) = borrows.get(cs.id) {
-            // SAFETY: shared reborrow of a DISTINCT child slot
-            // (parent ≠ child by tree acyclicity). No `&mut` to this
-            // slot is live: the recursive layout callback has not run
-            // yet, and `node_ref` covers only the parent's slot.
-            // Distinct slot reborrows have independent tags under
-            // Stacked / Tree Borrows.
-            let child_node: &RenderNode = unsafe { &*child_ptr.0 };
-            cs.offset = child_node.offset();
-            cs.needs_layout = child_node.needs_layout();
-            // Seed parent data from the child's persistent RenderState.
-            // This replaces the old behavior where parent_data was always
-            // None on the read path (rebuilt every frame).
-            cs.parent_data = child_node.parent_data().map(dyn_clone::clone_box);
-            if let Some(sliver_entry) = child_node.as_sliver() {
-                cs.sliver_constraints = sliver_entry.state().constraints().copied();
-                cs.sliver_geometry = sliver_entry.state().geometry();
-            }
-        }
-        #[cfg(any(test, feature = "testing"))]
-        borrows.seed_child_parent_data(cs.id, &mut cs.parent_data);
-    }
-
-    // Descendant-error tracking flag. Closure flips to `true` on any
-    // descendant `RenderError`; stage 6 below skips `clear_needs_layout`
-    // when set so the parent stays dirty for next-frame retry. Shared
-    // via `Arc<AtomicBool>` because the closure is `Send + Sync`
-    // (inherited from `LayoutChildCallback`'s bound).
-    let descendant_error_flag: std::sync::Arc<std::sync::atomic::AtomicBool> =
-        std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false));
-    let descendant_error_for_cb = std::sync::Arc::clone(&descendant_error_flag);
-
-    // Capture `&SubtreeBorrows` for the recursive callback. `&T` is
-    // `Send` iff `T: Sync`; `SubtreeBorrows: Sync` because its
-    // `HashMap<RenderId, NodePtr>` is Sync (NodePtr declares Sync via
-    // unsafe impl + RenderId is Sync) and the `PhantomData<&'tree mut ()>`
-    // is Sync too. So `&SubtreeBorrows: Send + Sync`, satisfying
-    // `LayoutChildCallback: Send + Sync`.
-    let borrows_for_cb: &SubtreeBorrows<'_> = borrows;
-    let descendant_error_for_sliver_cb = std::sync::Arc::clone(&descendant_error_flag);
-    let cb_owned = move |child_id: RenderId, child_constraints: BoxConstraints| -> Size {
-        // SAFETY: `borrows_for_cb` is alive (held by the outer
-        // layout_dirty_root stack frame for the entire walk). The
-        // recursive reborrow happens on `child_id`'s slot — distinct
-        // from the current `id`'s slot (tree acyclicity → parent ≠
-        // child). No two concurrent reborrows of the same NodePtr.
-        match unsafe { layout_subtree_borrowed(borrows_for_cb, child_id, child_constraints) } {
-            Ok(size) => size,
-            Err(err) => {
-                descendant_error_for_cb.store(true, std::sync::atomic::Ordering::Relaxed);
-                tracing::error!(
-                    parent = ?id,
-                    ?child_id,
-                    ?err,
-                    "layout_dirty_root: descendant layout failed; \
-                         returning Size::ZERO to caller's perform_layout. \
-                         Parent NEEDS_LAYOUT preserved for next-frame retry.",
-                );
-                Size::ZERO
-            }
-        }
-    };
-    let cb_ref: LayoutChildCallback<'_> = &cb_owned;
-
-    let baseline_cb_owned = move |child_id: RenderId, baseline: crate::traits::TextBaseline| {
-        borrows_for_cb.get(child_id).and_then(|child_ptr| {
-            // SAFETY: shared reborrow of a distinct child slot after its
-            // layout completed in this walk; no concurrent &mut to the slot.
-            let child_node: &RenderNode = unsafe { &*child_ptr.0 };
-            child_node
-                .as_box()
-                .and_then(|entry| entry.render_object().actual_baseline_raw(baseline))
-        })
-    };
-    let baseline_cb_ref: ActualBaselineChildCallback<'_> = &baseline_cb_owned;
-
-    // Sliver child callback: invoked when the Box parent calls
-    // `ctx.layout_sliver_child(index, sliver_constraints)`.  Uses the
-    // same `borrows_for_cb` pool and `descendant_error_flag` as the box
-    // callback — no extra pre-acquisition needed since sliver children
-    // are already in the pre-acquired `SubtreeBorrows` set.
-    let sliver_cb_owned = move |child_id: RenderId,
-                                sliver_constraints: SliverConstraints|
-          -> SliverGeometry {
-        // SAFETY: `borrows_for_cb` is alive for the entire walk (held
-        // by the `layout_dirty_root` stack frame). The reborrow targets
-        // `child_id`'s slot — distinct from the Box parent's slot
-        // (tree acyclicity) and from any concurrently-active Box child
-        // slot (LayoutCycleGuard blocks re-entry). No two concurrent
-        // reborrows of the same NodePtr.
-        match unsafe {
-            layout_sliver_subtree_borrowed(borrows_for_cb, child_id, sliver_constraints)
-        } {
-            Ok(geometry) => geometry,
-            Err(err) => {
-                descendant_error_for_sliver_cb.store(true, std::sync::atomic::Ordering::Relaxed);
-                tracing::error!(
-                    parent = ?id,
-                    ?child_id,
-                    ?err,
-                    "layout_dirty_root: sliver descendant layout failed; \
-                         returning SliverGeometry::ZERO to caller's perform_layout. \
-                         Parent NEEDS_LAYOUT preserved for next-frame retry.",
-                );
-                SliverGeometry::ZERO
-            }
-        }
-    };
-    let sliver_cb_ref: SliverLayoutChildCallback<'_> = &sliver_cb_owned;
-
-    // Construct the driver-side PARENT-DATA-ERASED context. The walk
-    // cannot name the parent's ParentData type (it holds dyn nodes);
-    // the typed blanket bridge reconstructs BoxLayoutCtx<T::Arity,
-    // T::ParentData> per node and lazily creates each child's
-    // parent-data slot with T::ParentData::default() — Flex/Stack and
-    // every other non-BoxParentData parent now lay out in production
-    // (the former ChildState<BoxParentData> hardcode panicked them in
-    // from_erased).
-    let mut ctx = crate::protocol::ErasedBoxLayoutCtx::new(
-        constraints,
-        &mut child_states,
-        &child_ids,
-        cb_ref,
-        baseline_cb_ref,
-        Some(sliver_cb_ref),
-    );
-    let erased: &mut dyn BoxLayoutCtxErased = &mut ctx;
-
-    // Invoke perform_layout_raw wrapped in catch_unwind (symmetric
-    // with the leaf path's layout_leaf_only — third-party panics
-    // surface as RenderError::Poisoned instead of unwinding out of
-    // layout_dirty_root). Capture debug_name BEFORE the &mut reborrow.
-    let debug_name = entry.render_object().debug_name();
-    let render_object = entry.render_object_mut();
-    let unwind_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-        render_object.perform_layout_raw(erased)
-    }));
-    let geometry = match unwind_result {
-        Ok(inner) => inner?,
-        Err(payload) => {
-            let msg = payload
-                .downcast_ref::<String>()
-                .map(String::as_str)
-                .or_else(|| payload.downcast_ref::<&'static str>().copied())
-                .unwrap_or("(non-string panic payload)");
-            tracing::error!(
-                render_object = debug_name,
-                panic_msg = msg,
-                "perform_layout panicked in non-leaf path — surfacing as \
-                     RenderError::Poisoned (symmetric with leaf-path \
-                     layout_leaf_only catch_unwind discipline)",
-            );
-            return Err(crate::error::RenderError::poisoned(debug_name, "layout"));
-        }
-    };
-
-    // State update on success path (mirrors
-    // RenderEntry::layout_leaf_only's post-perform_layout discipline).
-    // On the Err path above, state is intentionally unmodified so
-    // NEEDS_LAYOUT stays set for next-frame retry.
-
-    // Same protocol-generic geometry guards the leaf commit runs
-    // (RenderEntry::layout_leaf_only). Runtime validation happens before
-    // state commit; debug assertions then mirror Flutter's debug-only
-    // contract checks.
-    <BoxProtocol as Protocol>::validate_layout_output(debug_name, &constraints, &geometry)?;
-    <BoxProtocol as Protocol>::debug_assert_layout_output(&constraints, &geometry);
-
-    entry.state_mut().set_geometry(geometry);
-    entry.state_mut().set_constraints(constraints);
-
-    // Commit the offsets perform_layout wrote via `position_child`
-    // into each child's persisted `RenderState.offset`. The
-    // `ChildState` vec is a per-walk transient — without this commit
-    // every positioned offset dies with the stack frame and paint /
-    // hit-test (which read `RenderState.offset` as the authoritative
-    // child position) would place all children at the parent origin.
-    // Runs only on the parent-success path: on the Err / panic paths
-    // above, state stays unmodified so NEEDS_LAYOUT retry semantics
-    // hold. A descendant error does NOT skip the commit — the
-    // parent's perform_layout returned Ok, so its positioning
-    // decisions are valid regardless of a failed grandchild.
-    for cs in &child_states {
-        if let Some(child_ptr) = borrows.get(cs.id) {
-            // SAFETY: shared reborrow of a DISTINCT child slot
-            // (parent ≠ child by tree acyclicity). All recursive
-            // child borrows ended when perform_layout_raw returned;
-            // `entry`/`node_ref` cover only the parent's slot.
-            // `set_offset` is an atomic store through `&self`.
-            let child_node: &RenderNode = unsafe { &*child_ptr.0 };
-            child_node.set_offset(cs.offset);
-        }
-    }
-
-    // Bootstrap relayout boundary (U17). BoxProtocol runs the Flutter
-    // formula; SliverProtocol would be a no-op (not reachable on this
-    // path since we routed through as_box_mut).
-    let has_parent = entry.links().parent().is_some();
-    let sized_by_parent = entry.render_object().sized_by_parent();
-    <BoxProtocol as Protocol>::bootstrap_relayout_boundary(
-        entry.state(),
-        sized_by_parent,
-        has_parent,
-    );
-
-    // Only clear NEEDS_LAYOUT if the recursive callback observed no
-    // descendant failure. Preserves retry-next-frame semantics.
-    if !descendant_error_flag.load(std::sync::atomic::Ordering::Relaxed) {
-        entry.clear_needs_layout();
-    } else {
-        tracing::debug!(
-            parent = ?id,
-            "layout_dirty_root: a descendant errored during this walk; \
-                 keeping parent NEEDS_LAYOUT set for next-frame retry"
-        );
-    }
-
-    Ok(geometry)
-}
-
-/// Recursive Box intrinsic query over the pre-acquired layout subtree.
-///
-/// Used when a Sliver parent needs Flutter-style Box child intrinsics during
-/// layout. It shares the layout walk's [`SubtreeBorrows`] pool instead of
-/// re-entering `PipelineOwner`, preserving the same disjoint-slot discipline as
-/// `layout_subtree_borrowed`.
-unsafe fn box_intrinsic_query_borrowed<'tree>(
-    borrows: &SubtreeBorrows<'tree>,
-    id: RenderId,
-    dimension: crate::storage::IntrinsicDimension,
-    extent: f32,
-) -> crate::error::RenderResult<f32> {
-    ensure_stack(|| {
-        // SAFETY: forwarded from this wrapper; ensure_stack only changes stack
-        // placement, not borrow lifetimes or aliasing.
-        unsafe { box_intrinsic_query_borrowed_impl(borrows, id, dimension, extent) }
-    })
-}
-
-/// Body of [`box_intrinsic_query_borrowed`].
-///
-/// # Safety
-///
-/// Same contract as [`layout_subtree_borrowed`]: `borrows` must outlive this
-/// call and recursive child callbacks, and re-entry into an in-flight node is
-/// rejected by [`LayoutCycleGuard`] before any second mutable reborrow occurs.
-unsafe fn box_intrinsic_query_borrowed_impl<'tree>(
-    borrows: &SubtreeBorrows<'tree>,
-    id: RenderId,
-    dimension: crate::storage::IntrinsicDimension,
-    extent: f32,
-) -> crate::error::RenderResult<f32> {
-    let _cycle_guard = LayoutCycleGuard::enter(borrows, id)?;
-
-    let NodePtr(node_ptr) = match borrows.get(id) {
-        Some(np) => np,
-        None => return Err(crate::error::RenderError::NodeNotFound(id)),
-    };
-
-    // SAFETY: this is the only live reborrow of `id` in this intrinsic query
-    // frame. Recursive callbacks target child slots; cycle guard rejects
-    // attempts to revisit an in-flight ancestor.
-    let node_ref: &mut RenderNode = unsafe { &mut *node_ptr };
-    let node_protocol = node_ref.protocol_name();
-    let entry: &mut RenderEntry<BoxProtocol> = match node_ref.as_box_mut() {
-        Some(e) => e,
-        None => {
-            return Err(crate::error::RenderError::ProtocolMismatch {
-                node_protocol,
-                constraints_protocol: "box",
-            });
-        }
-    };
-
-    if let Some(hit) = entry
-        .state()
-        .layout_cache()
-        .peek_intrinsic(dimension, extent)
-    {
-        return Ok(hit);
-    }
-
-    let child_ids: Vec<RenderId> = entry.links().children().to_vec();
-    let mut child_err: Option<crate::error::RenderError> = None;
-    let value = {
-        let child_err = &mut child_err;
-        let mut child_query =
-            |index: usize, dim: crate::storage::IntrinsicDimension, ext: f32| -> f32 {
-                let Some(&child_id) = child_ids.get(index) else {
-                    child_err.get_or_insert(crate::error::RenderError::contract_violation(
-                        "sliver box child intrinsic query",
-                        "child index out of range for this node's children",
-                    ));
-                    return 0.0;
-                };
-                // SAFETY: the child query uses the same pre-acquired subtree
-                // and targets a child slot distinct from the current box node.
-                match unsafe { box_intrinsic_query_borrowed(borrows, child_id, dim, ext) } {
-                    Ok(value) => value,
-                    Err(err) => {
-                        child_err.get_or_insert(err);
-                        0.0
-                    }
-                }
-            };
-        let mut child_flex = |index: usize| -> i32 {
-            child_flex_from_seeds(
-                #[cfg(any(test, feature = "testing"))]
-                &borrows.parent_data_seeds,
-                #[cfg(not(any(test, feature = "testing")))]
-                &(),
-                &child_ids,
-                index,
-            )
-        };
-        entry.render_object().intrinsic_raw(
-            dimension,
-            extent,
-            child_ids.len(),
-            &mut child_query,
-            &mut child_flex,
-        )
-    };
-
-    if let Some(err) = child_err {
-        return Err(err);
-    }
-
-    entry
-        .state_mut()
-        .layout_cache_mut()
-        .insert_intrinsic(dimension, extent, value);
-    Ok(value)
-}
-
-// ============================================================================
-// Cross-protocol child layout: Box parent → Sliver child
-// ============================================================================
-//
-// `layout_sliver_subtree_borrowed` is the Sliver sibling of
-// `layout_subtree_borrowed`. It is called from the `sliver_cb_owned`
-// closure captured inside `layout_subtree_borrowed_impl`'s non-leaf path
-// when a Box parent calls `ctx.layout_sliver_child(index, constraints)`.
-//
-// Scope: sliver subtrees. Leaf nodes still delegate to
-// `RenderEntry::layout_leaf_only`; non-leaf slivers get an erased driver
-// context and may call `ctx.layout_child(...)` to lay out sliver children.
-//
-// The short-circuit clean-child optimisation from the Box path is omitted
-// here on purpose: sliver constraints change with scroll position on every
-// frame. Omitting the check is correct and saves a branch.
-
-/// Stack-probe wrapper for [`layout_sliver_subtree_borrowed_impl`].
-///
-/// # Safety
-///
-/// Same contract as [`layout_subtree_borrowed`]:
-/// 1. `borrows` must outlive every recursive invocation this helper
-///    triggers (it is held by the outer `layout_dirty_root` stack frame
-///    for the entire walk).
-/// 2. At any moment, no two concurrent reborrows of the SAME [`NodePtr`]
-///    exist. The `LayoutCycleGuard` enforces this by returning
-///    [`crate::error::RenderError::LayoutCycle`] on re-entry into a slot
-///    already in flight, preventing the otherwise-UB second Unique tag.
-unsafe fn layout_sliver_subtree_borrowed<'tree>(
-    borrows: &SubtreeBorrows<'tree>,
-    id: flui_foundation::RenderId,
-    constraints: SliverConstraints,
-) -> crate::error::RenderResult<SliverGeometry> {
-    ensure_stack(|| {
-        // SAFETY: identical contract, forwarded verbatim from this
-        // wrapper's own `# Safety` section; the stack-growth wrapper
-        // only relocates which memory the frames live in, never their
-        // borrow structure, lifetimes, or drop order.
-        unsafe { layout_sliver_subtree_borrowed_impl(borrows, id, constraints) }
-    })
-}
-
-/// Body of [`layout_sliver_subtree_borrowed`]; split out so every
-/// recursion level enters through the [`ensure_stack`] probe.
-///
-/// # Safety
-///
-/// Same contract as [`layout_sliver_subtree_borrowed`].
-unsafe fn layout_sliver_subtree_borrowed_impl<'tree>(
-    borrows: &SubtreeBorrows<'tree>,
-    id: flui_foundation::RenderId,
-    constraints: SliverConstraints,
-) -> crate::error::RenderResult<SliverGeometry> {
-    // U21 cycle guard: register `id` in currently_laying_out *before*
-    // any NodePtr reborrow. Same RAII discipline as the Box path —
-    // Drop runs on every exit including unwind so the set stays
-    // consistent across panics.
-    let _cycle_guard = LayoutCycleGuard::enter(borrows, id)?;
-
-    // Resolve id → NodePtr. Cross-thread access panics inside `get`.
-    let NodePtr(node_ptr) = match borrows.get(id) {
-        Some(np) => np,
-        None => return Err(crate::error::RenderError::NodeNotFound(id)),
-    };
-
-    // SAFETY: `node_ptr` aliases a `&mut RenderNode` that
-    // `SubtreeBorrows` holds disjointly from every other slot it
-    // covers. The reborrow below is the FIRST and ONLY reborrow of
-    // `id`'s slot in this call frame; no other live `&mut` to this
-    // slot exists — the Box parent's reborrow belongs to a DISTINCT
-    // slot (parent ≠ child by tree acyclicity), and the cycle guard
-    // above prevents re-entry into `id` from the same call stack.
-    // Distinct slot reborrows have independent Unique tags under
-    // Stacked / Tree Borrows — no aliasing.
-    let node_ref: &mut crate::storage::RenderNode = unsafe { &mut *node_ptr };
-
-    let node_protocol = node_ref.protocol_name();
-    let entry: &mut RenderEntry<SliverProtocol> = match node_ref.as_sliver_mut() {
-        Some(e) => e,
-        None => {
-            return Err(crate::error::RenderError::ProtocolMismatch {
-                node_protocol,
-                constraints_protocol: "Sliver",
-            });
-        }
-    };
-    let child_ids: Vec<RenderId> = entry.links().children().to_vec();
-
-    // No early-return here for the empty case: a lazy sliver (e.g.
-    // `RenderSliverListLazy`) starts with zero attached children and must
-    // call `build_and_layout_box_child` on its first frame to schedule
-    // initial builds via `pending_builds`. The `ErasedSliverLayoutCtx`
-    // path is always taken for slivers so that the context's
-    // `pending_builds` / `pending_removes` sinks are wired in — even when
-    // the current child list is empty.
-    let mut child_states: Vec<crate::protocol::ErasedSliverChildState> = child_ids
-        .iter()
-        .map(|&cid| crate::protocol::ErasedSliverChildState::new(cid))
-        .collect();
-
-    // Seed child offsets and parent-data from persisted `RenderNode` state.
-    //
-    // Offset: a sliver parent that does not call `position_child` on a later
-    // pass preserves its child's previous placement, matching Box semantics.
-    //
-    // Parent-data: the transient `ErasedSliverChildState.parent_data` starts
-    // as `None`; seed it from the node's persisted parent-data so that lazy
-    // sliver consumers (e.g. `RenderSliverListLazy`) can read the logical
-    // index installed by `apply_deferred_mutation` on the previous frame.
-    // `ParentData: DynClone` makes this a cheap heap clone (one `Box` per
-    // attached child; K is bounded by viewport/cache band, not by item count).
-    for cs in &mut child_states {
-        if let Some(child_ptr) = borrows.get(cs.id) {
-            // SAFETY: shared reborrow of a DISTINCT child slot
-            // (parent != child by tree acyclicity). No recursive child
-            // borrow has run yet in this frame, and `node_ref` covers only
-            // the current sliver parent slot.
-            let child_node: &crate::storage::RenderNode = unsafe { &*child_ptr.0 };
-            cs.offset = child_node.offset();
-            if let Some(pd) = child_node.parent_data() {
-                cs.parent_data = Some(dyn_clone::clone_box(pd));
-            }
-        }
-        #[cfg(any(test, feature = "testing"))]
-        borrows.seed_child_parent_data(cs.id, &mut cs.parent_data);
-    }
-
-    let descendant_error_flag: std::sync::Arc<std::sync::atomic::AtomicBool> =
-        std::sync::Arc::new(std::sync::atomic::AtomicBool::new(false));
-    let descendant_error_for_cb = std::sync::Arc::clone(&descendant_error_flag);
-    let descendant_error_for_box_cb = std::sync::Arc::clone(&descendant_error_flag);
-    let descendant_error_for_intrinsic_cb = std::sync::Arc::clone(&descendant_error_flag);
-    let borrows_for_cb: &SubtreeBorrows<'_> = borrows;
-
-    let cb_owned = move |child_id: RenderId,
-                         child_constraints: SliverConstraints|
-          -> SliverGeometry {
-        // SAFETY: `borrows_for_cb` is alive for the whole dirty-root
-        // walk, and this callback reborrows a child slot distinct from
-        // the current sliver node. `LayoutCycleGuard` rejects re-entry.
-        match unsafe { layout_sliver_subtree_borrowed(borrows_for_cb, child_id, child_constraints) }
-        {
-            Ok(geometry) => geometry,
-            Err(err) => {
-                descendant_error_for_cb.store(true, std::sync::atomic::Ordering::Relaxed);
-                tracing::error!(
-                    parent = ?id,
-                    ?child_id,
-                    ?err,
-                    "layout_dirty_root: sliver descendant layout failed; \
-                     returning SliverGeometry::ZERO to caller's perform_layout. \
-                     Parent NEEDS_LAYOUT preserved for next-frame retry.",
-                );
-                SliverGeometry::ZERO
-            }
-        }
-    };
-    let cb_ref: SliverChildLayoutCallback<'_> = &cb_owned;
-
-    let box_cb_owned = move |child_id: RenderId, child_constraints: BoxConstraints| -> Size {
-        // SAFETY: same subtree-borrow contract as the sliver child callback,
-        // but routed through the Box layout walk for Sliver -> Box adapters.
-        match unsafe { layout_subtree_borrowed(borrows_for_cb, child_id, child_constraints) } {
-            Ok(size) => size,
-            Err(err) => {
-                descendant_error_for_box_cb.store(true, std::sync::atomic::Ordering::Relaxed);
-                tracing::error!(
-                    parent = ?id,
-                    ?child_id,
-                    ?err,
-                    "layout_dirty_root: box descendant layout failed from sliver parent; \
-                     returning Size::ZERO to caller's perform_layout. \
-                     Parent NEEDS_LAYOUT preserved for next-frame retry.",
-                );
-                Size::ZERO
-            }
-        }
-    };
-    let box_cb_ref: crate::protocol::sliver_protocol::BoxChildLayoutCallback<'_> = &box_cb_owned;
-
-    let box_intrinsic_cb_owned = move |child_id: RenderId,
-                                       dimension: crate::storage::IntrinsicDimension,
-                                       extent: f32|
-          -> f32 {
-        // SAFETY: same subtree-borrow contract as the Sliver -> Box layout
-        // callback, but this read-only sizing query routes through the Box
-        // intrinsic bridge instead of perform_layout.
-        match unsafe { box_intrinsic_query_borrowed(borrows_for_cb, child_id, dimension, extent) } {
-            Ok(value) => value,
-            Err(err) => {
-                descendant_error_for_intrinsic_cb.store(true, std::sync::atomic::Ordering::Relaxed);
-                tracing::error!(
-                    parent = ?id,
-                    ?child_id,
-                    ?err,
-                    "layout_dirty_root: box intrinsic query failed from sliver parent; \
-                     returning 0.0 to caller's perform_layout. \
-                     Parent NEEDS_LAYOUT preserved for next-frame retry.",
-                );
-                0.0
-            }
-        }
-    };
-    let box_intrinsic_cb_ref: crate::protocol::sliver_protocol::BoxChildIntrinsicCallback<'_> =
-        &box_intrinsic_cb_owned;
-
-    let mut ctx = crate::protocol::ErasedSliverLayoutCtx::new(
-        constraints,
-        &mut child_states,
-        &child_ids,
-        cb_ref,
-        box_cb_ref,
-        box_intrinsic_cb_ref,
-        id,
-        &borrows.pending_builds,
-        &borrows.pending_removes,
-    );
-    let erased: &mut dyn SliverLayoutCtxErased = &mut ctx;
-
-    let debug_name = entry.render_object().debug_name();
-    let render_object = entry.render_object_mut();
-    let unwind_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-        render_object.perform_layout_raw(erased)
-    }));
-    let geometry = match unwind_result {
-        Ok(inner) => inner?,
-        Err(payload) => {
-            let msg = payload
-                .downcast_ref::<String>()
-                .map(String::as_str)
-                .or_else(|| payload.downcast_ref::<&'static str>().copied())
-                .unwrap_or("(non-string panic payload)");
-            tracing::error!(
-                render_object = debug_name,
-                panic_msg = msg,
-                "perform_layout panicked in non-leaf sliver path — surfacing as \
-                 RenderError::Poisoned",
-            );
-            return Err(crate::error::RenderError::poisoned(debug_name, "layout"));
-        }
-    };
-
-    <SliverProtocol as Protocol>::validate_layout_output(debug_name, &constraints, &geometry)?;
-    <SliverProtocol as Protocol>::debug_assert_layout_output(&constraints, &geometry);
-
-    entry.state_mut().set_geometry(geometry);
-    entry.state_mut().set_constraints(constraints);
-
-    // Commit child paint offsets produced by sliver parents
-    // (`RenderSliverPadding` et al.) into the same parent-relative
-    // offset slot that paint / hit-test already consult. The
-    // `ErasedSliverChildState` vec is per-walk transient; without
-    // this commit, child placement dies with the layout stack frame.
-    for cs in &child_states {
-        if let Some(child_ptr) = borrows.get(cs.id) {
-            // SAFETY: shared reborrow of a DISTINCT child slot
-            // (parent != child by tree acyclicity). All recursive
-            // child borrows ended when perform_layout_raw returned;
-            // `entry` / `node_ref` cover only the parent's slot.
-            let child_node: &crate::storage::RenderNode = unsafe { &*child_ptr.0 };
-            child_node.set_offset(cs.offset);
-        }
-    }
-
-    let has_parent = entry.links().parent().is_some();
-    let sized_by_parent = entry.render_object().sized_by_parent();
-    <SliverProtocol as Protocol>::bootstrap_relayout_boundary(
-        entry.state(),
-        sized_by_parent,
-        has_parent,
-    );
-
-    if !descendant_error_flag.load(std::sync::atomic::Ordering::Relaxed) {
-        entry.clear_needs_layout();
-    } else {
-        tracing::debug!(
-            parent = ?id,
-            "layout_dirty_root: a sliver descendant errored during this walk; \
-             keeping parent NEEDS_LAYOUT set for next-frame retry"
-        );
-    }
-
-    Ok(geometry)
 }
 
 // ============================================================================
@@ -4324,35 +3221,6 @@ where
 struct QuerySlot<'a> {
     node: Option<&'a mut crate::storage::RenderNode>,
     children: Vec<RenderId>,
-}
-
-#[cfg(any(test, feature = "testing"))]
-fn flex_factor_from_seed(seed: &ParentDataSeed) -> i32 {
-    match seed {
-        ParentDataSeed::Flex(data) => data.flex.unwrap_or(0).max(0),
-        _ => 0,
-    }
-}
-
-fn child_flex_from_seeds(
-    #[cfg(any(test, feature = "testing"))] parent_data_seeds: &FxHashMap<RenderId, ParentDataSeed>,
-    #[cfg(not(any(test, feature = "testing")))] _parent_data_seeds: &(),
-    children: &[RenderId],
-    index: usize,
-) -> i32 {
-    #[cfg(any(test, feature = "testing"))]
-    {
-        children
-            .get(index)
-            .and_then(|child_id| parent_data_seeds.get(child_id))
-            .map(flex_factor_from_seed)
-            .unwrap_or(0)
-    }
-    #[cfg(not(any(test, feature = "testing")))]
-    {
-        let _ = (children, index);
-        0
-    }
 }
 
 /// Recursive memoized intrinsic query over the take-out slot map.
